Some cleanup, removed reboot userspace post-restore option, manually get all packages instead of shelling out, int -> NSUInteger, adjusted list backup prompt wording, and changed a one to a zero

+++ Managers/IALRestoreManager.h
@@ -3,8 +3,8 @@
 @class IALGeneralManager;
 
 @interface IALRestoreManager : NSObject <NSURLSessionTaskDelegate, NSURLSessionDownloadDelegate>  {
-    int _expectedDownloads;
-    int _actualDownloads;
+    NSUInteger _expectedDownloads;
+    NSUInteger _actualDownloads;
 }
 @property (nonatomic, retain) IALGeneralManager *generalManager;
 -(void)restoreFromBackup:(NSString *)backupName ofType:(NSInteger)type;
